requirements.txt
Flask
python-pptx
gunicorn
Werkzeug
requests
typing